Removed StreamSocket.cs
